Homework #9
WorkManager

Реализовано обновление кэша фильмов с использованием WorkManager
Дата и время логгируется в файле cache/Log.txt

+++ app/src/main/java/com/example/superpupermegaproject/model/database/entities/MovieEntity.kt
@@ -17,7 +17,7 @@ data class MovieEntity(
         val movieID: Long,
 
         @ColumnInfo(name = moviesTableContract.COLUMN_NAME_POSITION)
-        val position: Int,
+        val position: Long,
 
         @ColumnInfo(name = moviesTableContract.COLUMN_NAME_TITLE)
         val title: String,
